Refactor menu forms into reusable components
